changed 1

+++ hash_tables/1-djb2.c
@@ -10,7 +10,7 @@
 
 unsigned long int hash_djb2(const unsigned char *str)
 {
-	unsigned int hash = 5381;
+	unsigned long int hash = 5381;
 	int c;
 
 	while ((c = *str++))
